add binds from my personal editor
Unfortunately neovim is still needed until I finish addressing a few
remaining issues in my personal editor (https://github.com/bd53/ate).

While it is fully functional and works as intended, there are still some
small tweaks and adjustments that need to be made for it to be fully complete.